requirements.txt
pymongo==3.9.0
flask==2.0.1
endesive==2.0.6
PyJWT==2.1.0
Werkzeug==2.0.1
python-dotenv==0.18.0
dnspython==2.1.0
pydantic==1.8.2
fastapi